CharTypesMap; GetAlphabetFor; debug output
Remove CharByCharType dictionary from CharTypesHelper in favor of just doing
reverse lookups with Linq on the CharTypeByChar dictionary (now renamed to
CharTypesMap) when we need a char and we have a CharType.

Write `GetAlphabetFor' method on CharTypesHelper so we can set the Alphabet
property on an FA by calling GetAlphabetFor(this).

Improve debug output formatting.